+++ packages/graphico/CHANGELOG.md
@@ -1,4 +1,11 @@
-# CHANGELOG
+# Changelog
+
+## 1.0.0
+
+- All layers are drawn onto a single canvas element
+- Support simple screen recording (no audio)
+- Updates to paint example
+- Add main README file
 
 ## 0.0.5